refactor(onboarding): merge provider gallery into AI setup step
The standalone providers step duplicated StepAISetup's provider selection
(Ollama appeared in both). Folded the gallery into the Cloud & Local branch:
picking a card installs it, makes it the active connection (baseUrl + default
model prefill), with manual baseUrl/key/model behind an advanced toggle. Removed
the extra step + progress slot. New ProviderGallery component is reusable.
