+++ changelogs/v2.6.15.md
@@ -2,7 +2,7 @@
 
 ### Added
 
-- **Onboarding now offers a one-tap AI provider gallery.** After the AI setup step, a new step shows a fixed-height two-per-row grid of community providers (OpenRouter, Ollama, Mistral and friends) — each card carries its own add state and prompts inline for an API key when one is needed, with keys stored in the OS keychain. It reuses the existing community provider registry, so the frame never jumps as providers load or install. You can still skip and add providers later from Settings → AI.
+- **Onboarding AI setup now offers a one-tap provider gallery.** The "Cloud & Local" branch of the AI setup step shows a fixed-height two-per-row grid of community providers (OpenRouter, Ollama, Mistral and friends) — each card carries its own add state and prompts inline for an API key when one is needed, with keys stored in the OS keychain. Picking a provider installs it and makes it the active connection (base URL + default model prefill); a "Manual endpoint / custom model" toggle keeps the raw baseUrl/key/model fields for anything not in the catalog. The frame never jumps as providers load or install, and you can always add more later from Settings → AI.
 
 ### Changed
